CRUD details complet

+++ app/Http/Controllers/Admin/PlanController.php
@@ -53,10 +53,19 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();   
+        $plan = $this->repository
+                            ->with('details')
+                            ->where('url', $url)
+                            ->first();   
 
         if (!$plan)
             return redirect()->back();
+
+        if ($plan->details->count() > 0) {
+            return redirect()
+                        ->back()
+                        ->with('error', 'Existem detalhes vinculados a esse plano, portanto para remover o plano é necessario deletar os detalhes');
+        }
         
         $plan->delete();
